+++ alicloud/resource_alicloud_ram_policy_test.go
@@ -160,6 +160,153 @@ func TestAccAlicloudRamPolicy_version_limit(t *testing.T) {
 
 }
 
+func TestAccAlicloudRamPolicy_withStstement(t *testing.T) {
+	var v ram.Policy
+	randInt := acctest.RandIntRange(1000000, 99999999)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+
+		// module name
+		IDRefreshName: "alicloud_ram_policy.policy",
+
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckRamPolicyDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccRamPolicyConfig_withStstement(randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description",""),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "version", "1"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "statement.#", "1"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "force", "false"),
+				),
+			},
+		},
+	})
+
+}
+
+func TestAccAlicloudRamPolicy_withDocument(t *testing.T) {
+	var v ram.Policy
+	randInt := acctest.RandIntRange(1000000, 99999999)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+
+		// module name
+		IDRefreshName: "alicloud_ram_policy.policy",
+
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckRamPolicyDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccRamPolicyConfig_withDocument(testPolicyTemplate1, randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description",""),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "version", "1"),
+					resource.TestCheckNoResourceAttr("alicloud_ram_policy.policy", "statement"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "force", "false"),
+				),
+			},
+		},
+	})
+
+}
+
+func TestAccAlicloudRamPolicy_reDocument(t *testing.T) {
+	var v ram.Policy
+	randInt := acctest.RandIntRange(1000000, 99999999)
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+
+		// module name
+		IDRefreshName: "alicloud_ram_policy.policy",
+
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckRamPolicyDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccRamPolicyConfig_reDocument(testPolicyTemplate1, randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description", "this is a policy test"),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "document", regexp.MustCompile("Allow*")),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "force", "true"),
+				),
+			},
+			{
+				Config: testAccRamPolicyConfig_reDocument(testPolicyTemplate2, randInt),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description", "this is a policy test"),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "document", regexp.MustCompile("Deny*")),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "force", "true"),
+				),
+			},
+		},
+	})
+
+}
+
+func TestAccAlicloudRamPolicy_reStatement(t *testing.T) {
+	var v ram.Policy
+
+	resource.Test(t, resource.TestCase{
+		PreCheck: func() {
+			testAccPreCheck(t)
+		},
+
+		// module name
+		IDRefreshName: "alicloud_ram_policy.policy",
+
+		Providers:    testAccProviders,
+		CheckDestroy: testAccCheckRamPolicyDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testAccRamPolicyConfig(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description", "this is a policy test"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "statement.#", "1"),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "version", "1"),
+				),
+			},
+			{
+				Config: testAccRamPolicyConfig_reStatement(acctest.RandIntRange(1000000, 99999999)),
+				Check: resource.ComposeTestCheckFunc(
+					testAccCheckRamPolicyExists("alicloud_ram_policy.policy", &v),
+					resource.TestMatchResourceAttr("alicloud_ram_policy.policy", "name", regexp.MustCompile("^tf-testAccRamPolicyConfig-*")),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "description", "this is a policy test"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "statement.#", "2"),
+					resource.TestCheckResourceAttrSet("alicloud_ram_policy.policy", "document"),
+					resource.TestCheckResourceAttr("alicloud_ram_policy.policy", "version", "1"),
+				),
+			},
+		},
+	})
+
+}
+
 func testAccCheckRamPolicyExists(n string, policy *ram.Policy) resource.TestCheckFunc {
 	return func(s *terraform.State) error {
 		rs, ok := s.RootModule().Resources[n]
@@ -253,3 +400,93 @@ func testAccRamPolicyConfig_version_limit(sequenceNO int) string {
 	  force = true
 	}`, sequenceNO)
 }
+
+func testAccRamPolicyConfig_withStstement(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_policy" "policy" {
+	  name = "tf-testAccRamPolicyConfig-%d"
+	  statement = [
+	    {
+	      effect = "Deny"
+	      action = [
+		"oss:ListObjects",
+		"oss:ListObjects"]
+	      resource = [
+		"acs:oss:*:*:mybucket",
+		"acs:oss:*:*:mybucket/*"]
+	    }]
+	}`, rand)
+}
+
+func testAccRamPolicyConfig_withDocument(policy string, rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_policy" "policy" {
+	  name = "tf-testAccRamPolicyConfig-%d"
+	  document = <<EOF
+      %s
+      EOF
+	}`, rand, policy)
+}
+
+func testAccRamPolicyConfig_reStatement(rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_policy" "policy" {
+	  name = "tf-testAccRamPolicyConfig-%d"
+	  statement = [
+	    {
+	      effect = "Deny"
+	      action = [
+		"oss:ListObjects",
+		"oss:ListObjects"]
+	      resource = [
+		"acs:oss:*:*:mybucket",
+		"acs:oss:*:*:mybucket/*"]
+	    },
+		{
+		  effect = "Allow"
+		  action = ["CreateInstance"]
+		  resource = ["*"]
+		}]
+	  description = "this is a policy test"
+	  force = true
+	}`, rand)
+}
+
+func testAccRamPolicyConfig_reDocument(policy string, rand int) string {
+	return fmt.Sprintf(`
+	resource "alicloud_ram_policy" "policy" {
+	  name = "tf-testAccRamPolicyConfig-%d"
+	  document = <<EOF
+      %s
+      EOF
+	  description = "this is a policy test"
+	  force = true
+	}`, rand, policy)
+}
+
+var testPolicyTemplate1 = `
+    {
+      "Version": "1",
+      "Statement": [
+        {
+          "Action": ["CreateInstance"],
+          "Resource": "*",
+          "Effect": "Allow"
+        }
+      ]
+    }
+`
+
+var testPolicyTemplate2 = `
+    {
+      "Version": "1",
+      "Statement": [
+        {
+          "Action": ["CreateInstance"],
+          "Resource": "*",
+          "Effect": "Deny"
+        }
+      ]
+    }
+`
+